+++ client/test/api_azure_cloud_credential_test.go
@@ -58,6 +58,18 @@ func Test_taikuncore_AzureCloudCredentialAPIService(t *testing.T) {
 
 	})
 
+	t.Run("Test AzureCloudCredentialAPIService AzureFileShareList", func(t *testing.T) {
+
+		t.Skip("skip test")  // remove to run test
+
+		resp, httpRes, err := apiClient.AzureCloudCredentialAPI.AzureFileShareList(context.Background()).Execute()
+
+		require.Nil(t, err)
+		require.NotNil(t, resp)
+		assert.Equal(t, 200, httpRes.StatusCode)
+
+	})
+
 	t.Run("Test AzureCloudCredentialAPIService AzureList", func(t *testing.T) {
 
 		t.Skip("skip test")  // remove to run test